ASF 2.0: Send correct user info from backend hooks
BZ: 166318

A recent patch corrected a deadlock by converting
ActivityManager.getCurrentUser() calls to
UserHandle.getCallingUserId().  Unfortunately
UserHandle.getCallingUserId() is not sufficient to accurately provide
user information to ASF, because these hooks don't actually execute
in a binder context.  This patch switches to using the "user" object
which is passed through the PackageManagerService.

When the user object contains a userId of -1 (which symbolically
represents "all users"), we provide a null UserInfo to ASF.  When the
user object is null, we provide the owner UserInfo.  In other cases,
we pass a UserInfo which is based on the user.

Change-Id: I231f22ff6e694abd531e541143660051076ef9b8
Category: feature differentiation
Domain: AOSP.Security-ASF
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Simmons, DavidX <davidx.l.simmons@intel.com>