Fix soft reboot while formatting
While formatting it's possible to cause a framework crash if
the format opperation is not sucessfull.

The 'getCode' function in NativeDaemonConnectorException can
cause a NullPointerException, since 'mEvent' can be null.

If an exception is thrown while formatting the thread will
try to create a Toast to notyfy the user of failure. This causes a
RuntimeException because a Toast cannot be shown from a background
thread.

Make sure 'getCode' will not thow a NullPointerException and also
create a handler so that the thread can send a message in order to
create a Toast.

Issue: APDEV-1219
Category: Aosp-Improvement
Domain: Aosp-storage
Origin: internal
Upstream-Candidate: Yes
Signed-off-by: Adriana Reus <adriana.reus@intel.com>
Change-Id: I216a384d1be8f516456cd6013916f6e4ea87637e