EM/THERMAL: Fix battery sensor detection logic
BZ: 168658

[Issue Description]
The ThermalService detects the thermal zone 'id' by
comparing the 'type' interface value under /sys/class/
thermal/thermal_zoneX/ and the sensor name configured
in thermal_sensor_config.xml. If it does not find, it
returns -1. This works well for sensors that have the
same name in XML and the sysfs 'type' interface.
[say step 1]

Unfortunately, the sysfs exposed by the battery sensor
carries different names for different platforms. Hence,
the XML uses "battery" whereas the sysfs might have
"[smb|max|bq|XXX]_battery" depending on the chip that is
used on that platform. So, as a special case, we detect
the battery sensor by comparing the exact string but by
checking whether the sysfs value "contains" the string
'battery'. [say step 2]

The current code, in case of failure in step 1, it
_always_ tries to find the battery zone id and assigns
it to all zones. Ideally, it should return with -1.

[Solution]
This patch fixes that by checking for the battery zone
only if the 'sensor name' configured in the XML is
"battery" (i.e only if it is a battery zone).

Category: device enablement
Domain: ThermalManagement-Common
Origin: internal
Upstream-Candidate: yes
Change-Id: I8ff1265b624f0ae82eb7baf007b4e854078614ca
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Sumeet Pawnikar <sumeet.r.pawnikar@intel.com>