gfx: Google fbo cache managment to improve perf
BZ: 163925

This is a temporary patch from IMG

The patch modifies the FboCache functionality in hwui
to not treat all FBO names equally. FBOs are cached
based on texture dimension; only matching texture
dimension FBOs are returned.

When the cache is full, instead of evicting based on LRU,
the current implementation frees the current (fresher)
FBO. This means the optimization will not be effective
if there are more than 16 layer FBOs. This doesn't seem
to be a problem for either Launcher3 or System UI.

Change-Id: Iab8ed8f1fdd5e0393dea3bd964ea7c05b11ce26d
Category: aosp improvement
Domain: Graphics-3D
Origin: 3rd-party IMG
Upstream-Candidate: no, not-aosp
Signed-off-by: insoowoo <insoo.woo@intel.com>