Wifi: Adjust dhcp range to the set ip address and netmask
BZ: 173206

Dhcp range for hotspot was hard coded because
before the user did not have the opprtunity to manually
set the hotspot ip address. Now that this is possible,
the dhcp range has to be adjusted according to the ip address
and netmask of the hostpot.

Change-Id: I48ecc230ee2062f36a5884d27bce9e2f259f4611
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>