wifi: do not call native layer from service
BZ: 175135

Currently the interfaces setSafeChannel and setRTCoexMode will
directly call the native layer (i.e. the supplicant) without
checking WifiStateMachine state. This may cause problem if the
state machine is currently stopping the supplicant.

This patch redirects native call in service to message for WifiStateMachine.
If the StateMachine is in a "valid" state, it calls the native layer
otherwise it discards the command. This implies that supplicant
return cannot be obtained synchronously, but since it is not really
used (no action is done if supplicant command failed), assume it is
not important.

Change-Id: I03aae8b350cab3cd4f2dc0d0833f542ffb990c43
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>