ASF 2.0: Resolve null pointer exception
BZ: 181797

There is at least one instance where sendBroadcastIntentEvent() is
called where the BroadcastRecord has not receivers, the
BroadcastRecord receivers method is null.  A check was added so that
the null pointer is not dereferenced.  When there are no receivers the
event is not sent because there will be no receivers to remove.

The ASF code added to scheduleBroadcastsLocked() is now guarded with
INTEL_FEATURE_ASF.

Change-Id: I31d8dffd691513497614f629bd5cfc9980cf36cc
Category: feature differentiation
Domain: AOSP.Security-ASF
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Norivn Shuster <norvin.shuster@intel.com>