EM/THERMAL: Add Emulated Temperature Event observer
BZ: 174693

This patch adds an UEvent observer to ThermalManager,
to listen to UEvents from kernel Thermal subsystem,
with event type as "2".

The kernel sends this category of UEvents when the
user/tools request to read the temperature of a zone
calibrated (using some Math) in the user space.
The Thermal Service responds to this UEvent by writing
the calibrated temperature to the 'emul_temp'
sysfs node under any thermal_zoneX.

Make the WriteToFile JNI method open files as
O_WRONLY.

DebounceInterval is supported only for Polling mode.
The HW/driver should take care of hysteresis for
interrupt mode.

For sensors that work on Interrupt mode as well as
have slope/intercept, the zone thresholds are
calibrated once during intialization; and the
calibrated thresholds are used from thereon.

Category: device enablement
Domain: ThermalManagement-Common
Origin: internal
Upstream-Candidate: yes
Change-Id: I1fe49f900376bc3c36ddd018a00f3df995091795
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>