[VIBRATOR] Set minimum vibration duration for vibration patterns
BZ: 190297

Currently, patterns can vibrate for a very short duration.
A minimum vibration duration is applied but only for single
vibrations.

Applying a minimum duration of the vibrations for patterns.
In case of a slow response of the vibrator, it will allow
to set a minimum value for the vibrations to be felt correctly by
the user.

Change-Id: I229434b582fcfb37d511e5834355b9efafe8d6c8
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>