Shutdown Time: Fix the issue of shutdown time is too high
Shutdown radio timeout value was set to 32 seconds.  Changed to 12 seconds
to equal OTC source code.  Also added check to see if IsVoiceCapable is
true or false.  If false, no need to wait for Phone Service to become false
(it never does).  This reduces the shutdown time on power button press to
approximately one second.

Change-Id: Ibf07dbb3400aad313f970aacd79575cf4a7a7483
Signed-off-by: sahashau <shaunak.saha@intel.com>
Signed-off-by: vvinjamu <venkateswarlu.v.vinjamuri@intel.com>
Reviewed-on: https://android.intel.com/188764
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>