Add hook patch for event processing before notify Motion event. And also fix the system server sometimes crash issue.
BZ: 591

Event processing library(named as touch filter) is used for reducing
and purifying the touch event, it would help to decrease the
power consumption. From the test on various devices and application,
it achieves 0.3% ~ 20% around system power reduction
when application running.

This patch is used to invoke event processing library.
When "libeventprocessing.so" library does exist under the "/system/lib",
it would be invoked to pre-handle events before these events are notified.