Telephony: add a check of network type in reassessPidDns
BZ: 198560

A NullPointerException occurred in the reassessPidDns()
function when IMS is disabled.

This exception is due to the fact that we don't check in
this function if the network type is valid or not before
to used it. In our case the IMS network types are not
valid and so not initialized.

This patch fixes the issue by adding a check of the network
type in the reassessPidDns() function before to used it.

Change-Id: I3f3f833e457a9015d26d5a85682119cf6ec63285
Category: feature differentiation
Domain: Telephony-IMS-IR92
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>