Keyguard: Drop-down list blocked after flight mode on/off with SIM card lock
BZ: 193209

When the SIM card lock is set to ON and after a reboot, if the flight mode
is set to ON then OFF, the drop-down list remains blocked.
Condition to disable the drop-down list was set so the feature is disabled
only if the secure screen lock is activated not also when the SIM card PIN
is requested.

Change-Id: Ia6e32e7b0843aa2f4b1da7a6d0c1f674b172c447
Category: aosp improvement
Domain: AOSP-Framework-GUI
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Sergiu Bratu <sergiux.bratu@intel.com>