Fix another crash in widget pager
Notes from Mike B:
Original patch came from Amith Dsouza (amithds@google.com), but we
believe there is an issue in where the currentPage != null check is, so
I have expanded the protected section. This patch may be replaced by
Google in a future update.
This is part of the 2014-06 Android Partner Security Bulletin.

Issue: GMIN-1552
Change-Id: I0d73e26a94fd20fc945fe186f80a336a3956f732
Signed-off-by: Michael Bergeron <michael.bergeron@intel.com>
Reviewed-on: https://android.intel.com/207871
Reviewed-by: Phillips, Scott D <scott.d.phillips@intel.com>
Reviewed-by: Demeter, Michael <michael.demeter@intel.com>