EM/Thermal: Revert airplane mode status before shutdown
BZ: 168834

If airplane mode is turned on by thermal service, it should
be turned off before shutdown. Else, the device will be in
airplane mode on the next boot which might be not be a good
user experience.
This patch checks if airplane mode is turned on by thermal
and if yes, turns it off before shutdown.

Change-Id: If48fcd4edd56a659fde4f9537b4479f29ca62bab
Category: device enablement
Domain: ThermalManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>