[PATCH 1/2] framework: show status as early as possible when power on
borqsbtx: 61372

rootcause: when power on, main thread is blooded with msg and blocked
sometimes, so use binder thread to make statusbar and navigation bar shown ASAP

Borig-Change-Id: Ia948fb772ee27423bc6dd88e03f4c02a0607f658
Reviewed-on: http://192.168.1.79/gerrit/7899
Reviewed-by: b023 <b023@borqs.com>

[PATCH 2/2] Gallery2: "com.android.wallpapercropper" crash when editing bmp & gif

borqsbtx: 0061392

rootcause: If the x, y, width, height values are outside of the dimensions of the source bitmap, it's failed to create bitmap.
When x < 0 or y < 0, make x = 0 or y = 0 and make width and height equal dimensions of the source bitmap to create bitmap.

Change-Id: I8059ee1b6e29a7eb585fde3e391cd65805946124
Borig-Change-Id: Ie71133049118aea62d75afd0fe89b92c661b9566
Reviewed-on: http://192.168.1.79/gerrit/7909
Reviewed-by: b023 <b023@borqs.com>