AlarmManagerService: add trace to know what intent triggers
This adds a trace to know what intent triggered the alarm,
to know which application is waking up the platform and why
so that we can quickly identify the culprit from logs.
We print only alarms of type _WAKEUP.
This is an old feature already present on main thanks to :
https://android.intel.com/#/c/145508/
but the original patch has been reworked to be simpler.

Change-Id: I92bc9c5d63aaba340ad086b0b34f92b7306a9545
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1539
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com:443/252084