SystemUI: Fix brightness slider movement with keyboard
Issue: BMDA-883

Once the brightness value has been set in the UI, the value is used
to update the System Settings after which the slider on the UI gets
updated. This is done on a separate thread through an AsyncTask.

When the brightness value is changed from the UI rapidly (either
through the keyboard or the mouse, moving the slider in very small
steps quickly), the lag between saving the brightness value and
updating the slider on the UI causes the slider to be in perpetual
motion. Introducing a small delay of 250ms after calling the
AsyncTask, covers for this lag.

This delay does not affect the user experience while using touch
or the mouse as, in these cases, the System Settings is updated
only at the final position at which the slider is released and not
at every point in between as is the case with the keyboard.

Change-Id: Ic9a7ca72fdb75e55208f75bbc6f453066ffa24dd
Signed-off-by: Sadhana Rajendran <sadhanax.rajendran@intel.com>
Reviewed-on: https://android.intel.com/259822
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Venkatesh Murthy, HemanthX <hemanthx.venkatesh.murthy@intel.com>
Reviewed-by: Chitgope, Raghuveer <raghuveer.chitgope@intel.com>
Tested-by: Chitgope, Raghuveer <raghuveer.chitgope@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>