[REVERT ME] framework: workaround to fetch the correct abi for the bundled-apps with cluster installation
This __workaround__ aims to fetch the abi of bundled-apps installed
in cluster way. The PMS will try to check whether the arm libs are
existing for this app. If yes, this abi will be set as that of the
bundled app.
NOTE: The bundled-apps with monolithic installation are not the
targets of this commit.

The patch will be removed when x86 verson's GMS is ready

Change-Id: I6fa750623bffe8f0203871b783ea5eb48752816c
Signed-off-by: jgu21 <jinghui.gu@intel.com>