[CWS_WIDI_UX] Adapter reconnection upon AP just connected on another frequency
Since some WiFi chip has some limitation to offer high video quality while used
in muli-roles and multi-channels, the best solution is to use the same channel as
the access point when running in multi-roles.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can imagagine that
user first connects to video adapter and then connects to an access point. In
such situation, there is a high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Change-Id: Ia3f903b1a5b96e7af4db22c271a3764d35812dc4
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3723
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: https://android.intel.com:443/283644