Security:Android KK must support secure adb as CDD
Issue: BMDA-1085

Decoupled ADB debugging manager from UsbDeviceManager

Till now, the ADB debugging manager (also named UsbDebuggingManager) is
created from UsbDeviceManager since the debuging on most Android devices
is made through USB cable.
But, the adbd proess also listens on an IP port (which can be configured
or the default port 5555 can be used).
If we enable the secure feature in adbd, the back-end of this feature is
UsbDebuggingManager which handles the authorization of the connecting
host. But, UsbDebuggingManager only handles USB connections.
So, since adbd also listens on an IP port, we need to decouple
UsbDebuggingManger from USB and make it generic: DebuggingManager. Thus,
we will be able to make use of the secure feature in adbd also when
using it over IP, not just over USB.

This patch, does the followings:
- removes UsbDebuggingManager.java and creates a new service:
DebuggingService.java. This new service does exactly the same thing as
the old UsbDebuggingManager.java, but it will run as a system service,
having a permanent connection with adbd.
- The adb debugging controls are removed from UsbService and are put
into a new interface: IDebuggingManager, implemented by
DebuggingService.
- SystemServer is modified to add the new service: DebuggingService
- applications like SystemUI and Settings will be modify to use the new
service DebuggingService instead of UsbService when handling ADB debug
options (like: enable/disable and authorize). Settings will be updated
in a separate patch.

For: APDEV-896
Category: AOSP Improvement
Domain: AOSP-debugging
Origin: internal
Upstream-Candidate: Yes

Signed-off-by: Robert Chiras <robert.chiras@intel.com>

Conflicts:
	services/java/com/android/server/usb/UsbDeviceManager.java

Change-Id: I86403e6ec60a7abcbb97672c2b0af376d0fa0822
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/188425
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Krishnamoorthy, Jagadish <jagadish.krishnamoorthy@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>

Change-Id: Ic0db49c919ca6bba2912fb57d5f0d5a1b258cfc5
Signed-off-by: Sadhana Rajendran <sadhanax.rajendran@intel.com>
Reviewed-on: https://android.intel.com/296630
Reviewed-by: Chitgope, Raghuveer <raghuveer.chitgope@intel.com>
Tested-by: Chitgope, Raghuveer <raghuveer.chitgope@intel.com>
Tested-by: Vinjamuri, Venkateswarlu V <venkateswarlu.v.vinjamuri@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>