Fix Timed OTA Deferral policy.
Between R80 and R106 a patch from Google was added to frameworks.
This patch switched from using the list 'pkgs' to 'sortedPkgs' as
the primary list used for their OTA policy and our OTA policy still used
'pkgs' which could be empty. If empty, our OTA policy would not
continue. This patch uses 'sortedPkgs' as the primary list to create a
final list of apps to be compiled.

Issue: CAR-852

Change-Id: I88acbd6f58413fed54b2dce2b4c749d218f747e4
Category: device enablement
Domain: AOSP
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: James C Scott <james.c.scott@intel.com>