WFD: Set GO intent to 14
BZ: 343

MultiRole is requested on Sofia 3G platform.
Since Sofia 3G only supports single channel,
it is needed to force p2p frequency to wlan
frequency when it is working in STA + P2P mode.
That request DUT to behave as GO.

This patch is to maximize the chance to become
GO to choose the p2p operating channel.

Change-Id: I917b16aab6928884c6d3381b83ffe7cdeb85f159
Signed-off-by: junmingh <junming.hong@intel.com>