[PATCH 2/2] Rebase patches to L WFD: Set GO intent to 14
BZ: 343

MultiRole is requested on Sofia 3G platform.
Since Sofia 3G only supports single channel,
it is needed to force p2p frequency to wlan
frequency when it is working in STA + P2P mode.
That request DUT to behave as GO.

This patch is to maximize the chance to become
GO to choose the p2p operating channel.

Change-Id: Ia6e86063b9419882a379fea118bf680c94240857
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-12868
Signed-off-by: junmingh <junming.hong@intel.com>
Signed-off-by: Fei Jiang <fei.jiang@intel.com>