update ConnectivitySer wakeup alarm
BZ: 575

ConnnectiviySer sets a wakeup alarm, by default,
it's interval is 12 minutes which intends to
track traffic quality data for WiFi and mobile networks,
and the quality is tracked based on incidental traffic,
and not on specific measurements.
However, this will cause power consumption high
since it always wakes up system, so better to change the alarm interval.

This patch will do:
1. Skip alarm setting for flight mode with wifi off;
2. Update alarm interval.

Change-Id: I1f40dfa0e3844861d28e4432d9bdad4726dd5efe
Signed-off-by: Huajun Li <huajun.li@intel.com>