[PATCH] [PATCH] merge x86-missing patch for app issues
BZ: 674

ISV .so means the APK's private, main logical. It is opponent with "3rd
party .so". 3rd party .so is an integral partial. It is an engine or a
decoder or Ads who modifies JAVA or encryptor who modifies native code
or else. Anyway, 3rd party .so missing doesn’t means that ISV forgets it
and in other side 3rd party .so may be integrated by tools instead of
ISV.

We are looking for the "behave" which can show us the APK tried to do
something on x86 porting. ISV .so is the best signal.

The algorithm depends on "ThirdPartySO" list. We will be fully "x86
aggressive" if "ThirdPartySO" is empty.

"OEMWhiteList" list is a shortcut. Any apk in "OEMWhiteList" will be
installed with lib/x86/*.so.

Change-Id: I05cf39b1ee151772f6a4a68118de1c821a08d59e